install: list contents of HOMEBREW_PREFIX if abort
If we bail out because HOMEBREW_PREFIX isn't empty, let users know what is
present in that directory. Most users seemingly only run `ls` which doesn't
show hidden files/folders without extra commands, and in many cases when the
install is aborted `.git` is left with contents which causes an abort since
we actively check that directory.

Closes https://github.com/Homebrew/homebrew/issues/45692.
Closes https://github.com/Homebrew/homebrew/issues/45733.