install: trim El Cap abort
We haven't seen this for a fair while since Apple pushed further updates to the
SIP mechanism and believe it to be far less of an issue than it was previously.

If you encounter further failure, please file an Issue and we'll be happy to
help! At the moment though this advice is making people think there _will_ be an
issue at some point.