GH#637 install.sh fail when the umask is not 0022
HOMEBREW_PREFIX was created using mkdir -p which respects the user
umask, even when using sudo causing the wrong permissions to be set
when installing into /home/linuxbrew.

This fixes the issue by replacing mkdir with install creating the
missing parents with the specified permissions. To make it work on
linux and macos remember to only use single letter flags.