Use `/home/${USER}` instead of `${HOME}`
While trying to automate installing linuxbrew in an ansible playbook I
noticed that all went well, apart from creating the `.cache` directory
in the user's home.

This is because of multiple reasons:
1. To run in ansible I need it to `export NONINTERACTIVE=t`
2. To be able to execute it with sudo privileges through the ansible
   `become` functionality I need to already be root, and use sudo to
   switch back to the oriningal user.
3. It requires `-E` to preserve non-interactive
4. Because of this `-u` switches `${USER}` correctly but preserves
   `${HOME}` from `root`.

The workaround was to `export HOME=/home/{{ ansible_user_id }}` before
execution, but honestly this should not be needed.

For reference, this is the ansible task that I wrote:

```yml
- name: Install homebrew
  ansible.builtin.shell: |
    set -eo pipefail

    export NONINTERACTIVE=1
    export HOME=/home/{{ ansible_user_id }}

    sudo -E -u {{ ansible_user_id }} `pwd`/install.sh
  args:
    chdir: '{{ homebrew_registered_repository_location.stdout }}'
  when: not homebrew_already_installed.stat.exists
  become: true
```

Unfortunately I could not find a better way to be able to automate this
using ansible, reason why I'm proposing this change. If we replace the
usage of `${HOME}` with `/home/${USER}` it should be identical to the
current behaviour.