install.sh: ensure `master` is used for `git init`
This is not the case when `init.defaultBranch` is set to something else,
like `main`. And in that case, if `HOMEBREW_DEVELOPER` is also set, the
`Homebrew/brew` repository and `homebrew/core` (if tapped) will stay on
that branch instead of `master` (and won't be updated anymore since
there will be no tracking information available).

Also, for consistency, replace the short option `-q` with `--quiet`, the
long one.

Signed-off-by: Ruoyu Zhong <zhongruoyu@outlook.com>