install: add fish post install instructions
This adds the correct config file for the fish shell needed to load
the brew environment. Fish ends up using the `~/.config/fish/config.fish`
file during startup to load environment variables and scripts most
of the time. This is also used by asdf and ghcup for instance.

This was brought up in discussion here:
- https://github.com/orgs/Homebrew/discussions/4412

More info:
- https://fishshell.com/docs/current/index.html#configuration