Make some git invocations quiet
Currently, all the branches and tags are printed out by `git fetch`,
which I believe our end users are not concerned about.

This change suppresses some of the Git outputs with the `--quiet`
option, so no branches or tags are printed out. As the progress bar can
still be useful, `--progress` is used to keep it enabled.