Adding sudoless mode incorporating suggestions
This is a larger change but it:
1. Has a check to make sure all files in the Homebrew prefix are writable
2. Specifically handles CLT installation
3. Cleanly separates it out into its own function

Uses running `chmod` as a check. If it fails then something isn't writable.
Bonus is it also is a permissions repair.

For the CLT install with macOS sudoless mode, the first method will fail, so within the fallback we check for macOS_sudoless_mode and if we're using it then `abort` with the message about manual installation, instead of running the GUI pop up manual installation.

This still retains the original goal of installing in the official prefixes without sudo, so it doesn't result in an "Alternative (unsupported) installation".